Ko-AD/Cursor/python/requirements.docker.txt
audioop-lts>=0.2.0
pydub
gTTS
ffmpeg-python
pandas
google-genai
google-generativeai
yt-dlp
<<<<<<< Updated upstream
<<<<<<< Updated upstream
moviepy

# === TTS with Coqui XTTS ===
coqui-tts>=0.27.0
numpy>=1.26.0
hangul_romanize>=0.1.0  # Required for Korean TTS

# === Audio Processing for Ducking ===
pyloudnorm>=0.1.0
scipy>=1.13.0
torchaudio>=2.0.0
torch>=2.0.0
librosa>=0.11.0  # Audio resampling for VAD (avoids torchcodec issues)
soundfile>=0.12.0  # Audio file I/O (used by librosa and pydub)

# === Silero VAD (loaded via torch.hub, but torch is required) ===
# Note: Silero VAD model is automatically downloaded when first used
=======
moviepy
>>>>>>> Stashed changes
=======
moviepy
>>>>>>> Stashed changes
